css updates

+++ js/background.js
@@ -6,10 +6,12 @@ function startRequest() {
   window.setTimeout(startRequest, 5000);
 }
  
-localStorage.clear();
-SetInitalOption("GWNRT.RequestInterval", 60000);
-SetInitalOption("GWNRT.BackgroundTabs", false);
-SetInitalOption("GWNRT.error", null);
+if (firstRequest) {
+  localStorage.clear();
+  SetInitalOption("GWNRT.RequestInterval", 60000);
+  SetInitalOption("GWNRT.BackgroundTabs", false);
+  SetInitalOption("GWNRT.error", null);
+}
 
 
 startRequest();
